Added expetion handler and unique checker for patient and doctor
Modified services and modified unique username added exeption checker for them

+++ ezderm-service/src/main/java/com/ezderm/solution/service/impl/DoctorServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 import com.ezderm.solution.dao.DoctorDao;
 import com.ezderm.solution.dto.DoctorDto;
 import com.ezderm.solution.exception.DuplicateUsernameException;
+import com.ezderm.solution.exception.NotPermittedException;
 import com.ezderm.solution.exception.ResourceNotFoundException;
 import com.ezderm.solution.model.Doctor;
 import com.ezderm.solution.service.DoctorService;
@@ -19,43 +20,48 @@ import lombok.extern.log4j.Log4j2;
 @RequiredArgsConstructor
 @Log4j2
 public class DoctorServiceImpl implements DoctorService {
-	
+
 	private final DoctorDao doctorDao;
-	
 
 	@Override
 	@Transactional
-	public DoctorDto saveDoctor(DoctorDto doctorDto,String userName) {
-		if(this.doctorDao.existsByUsername(userName)) {
+	public DoctorDto saveDoctor(DoctorDto doctorDto, String userName) {
+		if (this.doctorDao.existsByUsername(userName)) {
 			throw new DuplicateUsernameException(" Username " + userName + " Already Exsist ");
 		}
-		
-		Optional<Doctor> existingDoctorOpt = this.doctorDao.findById(doctorDto.getId());
-	    Doctor doctor;
-	    if (existingDoctorOpt.isPresent()) {
-	        doctor = existingDoctorOpt.get();
-	        log.info("Updating existing doctor with ID: {}", doctor.getId());
-	    } else {
-	        doctor = new Doctor();
-	        log.info("Creating new doctor...");
-	    }
-	    doctor.setFirstName(doctorDto.getFirstName());
-	    doctor.setLastName(doctorDto.getLastName());
-	    doctor.setUsername(userName);
-	    doctor = this.doctorDao.save(doctor);
-	    log.info("Saved doctor with ID: {}", doctor.getId());	    
-	    return new DoctorDto(doctor.getId(),doctor.getFirstName(),doctor.getLastName());
-	    
-	}
 
+		Optional<Doctor> existingDoctorOpt = this.doctorDao.findById(doctorDto.getId());
+		Doctor doctor;
+		if (existingDoctorOpt.isPresent()) {
+			doctor = existingDoctorOpt.get();
+			log.info("Updating existing doctor with ID: {}", doctor.getId());
+		} else {
+			doctor = new Doctor();
+			log.info("Creating new doctor...");
+		}
+		doctor.setFirstName(doctorDto.getFirstName());
+		doctor.setLastName(doctorDto.getLastName());
+		doctor.setUsername(userName);
+		doctor.setUuid(doctorDto.getUuid());
+		doctor = this.doctorDao.save(doctor);
+		log.info("Saved doctor with ID: {}", doctor.getId());
+		return new DoctorDto(doctor.getId(), doctor.getUuid(), doctor.getFirstName(), doctor.getLastName());
+
+	}
 
 	@Override
 	@Transactional
-	public void deleteDoctor(Long doctorId) {
-		 Doctor doctor = this.doctorDao.findById(doctorId)
-			        .orElseThrow(() -> new ResourceNotFoundException("Doctor with ID " + doctorId + " not found"));
-		 log.info("Deleting doctor with ID: {}",doctorId);
-		 this.doctorDao.delete(doctor);
+	public DoctorDto deleteDoctor(String uuid, String userName) {
+		Doctor doctor = doctorDao.findByUuid(uuid);
+		if (doctor == null) {
+			throw new ResourceNotFoundException("Doctor with UUID " + uuid + " not found");
+		}
+		if (!doctor.getUsername().equals(userName)) {
+	        throw new NotPermittedException("User " + userName + " is not authorized to delete this doctor");
+	    }	
+		log.info("Deleting doctor with ID: {}", uuid);
+		this.doctorDao.delete(doctor);
+		return new DoctorDto(doctor.getId(),doctor.getUuid(),doctor.getFirstName(),doctor.getLastName());
 	}
 
 }
